refactor(tests): reclassify mock-based integration tests as unit tests
Move 6 fully-mocked integration test files to tests/unit/ and remove
@pytest.mark.integration markers so they run without --integration flag.
Extract shared helpers and inline fixture data to reduce duplication.

Moved files:
- test_authentication.py -> tests/unit/auth/
- test_cross_service.py -> tests/unit/servers/
- test_mcp_protocol.py -> tests/unit/servers/
- test_transport_lifecycle.py -> tests/unit/servers/
- test_stdin_monitoring_fix.py -> tests/unit/
- test_proxy.py -> tests/unit/utils/

Additional cleanup:
- Extract setup_api3_passthrough_mocks() to tests/utils/mocks.py
- Move inline data dicts from jira/confluence conftest to fixtures
- Remove unused jira_comments_data fixture from models/conftest.py
- Fix pre-existing test bugs exposed by removing integration skip:
  timeout=75 in auth assertions, OAuth precedence over basic auth,
  ASGI middleware interface (dispatch -> __call__), MagicMock
  auto-attribute creation causing false service availability
