+++ tests/unit/utils/test_proxy.py
@@ -1,5 +1,5 @@
 """
-Integration tests for proxy handling in Jira and Confluence clients (mocked requests).
+Unit tests for proxy handling in Jira and Confluence clients (mocked requests).
 """
 
 import os
@@ -16,7 +16,6 @@ from tests.utils.base import BaseAuthTest
 from tests.utils.mocks import MockEnvironment
 
 
-@pytest.mark.integration
 def test_jira_client_passes_proxies_to_requests(monkeypatch):
     """Test that JiraClient passes proxies to requests.Session.request."""
     mock_jira = MagicMock()
@@ -48,7 +47,6 @@ def test_jira_client_passes_proxies_to_requests(monkeypatch):
     assert mock_session.proxies["socks"] == "socks5://user:pass@proxy:1080"
 
 
-@pytest.mark.integration
 def test_confluence_client_passes_proxies_to_requests(monkeypatch):
     """Test that ConfluenceClient passes proxies to requests.Session.request."""
     mock_confluence = MagicMock()
@@ -87,7 +85,6 @@ def test_confluence_client_passes_proxies_to_requests(monkeypatch):
     assert mock_session.proxies["socks"] == "socks5://user:pass@proxy:1080"
 
 
-@pytest.mark.integration
 def test_jira_client_no_proxy_env(monkeypatch):
     """Test that JiraClient sets NO_PROXY env var and requests to excluded hosts bypass proxy."""
     mock_jira = MagicMock()
@@ -113,7 +110,6 @@ def test_jira_client_no_proxy_env(monkeypatch):
 class TestProxyConfigurationEnhanced(BaseAuthTest):
     """Enhanced proxy configuration tests using test utilities."""
 
-    @pytest.mark.integration
     def test_proxy_configuration_from_environment(self):
         """Test proxy configuration loaded from environment variables."""
         with MockEnvironment.basic_auth_env() as env_vars:
@@ -138,7 +134,6 @@ class TestProxyConfigurationEnhanced(BaseAuthTest):
                 assert confluence_config.https_proxy == "https://proxy.company.com:8443"
                 assert confluence_config.no_proxy == "*.internal.com,localhost"
 
-    @pytest.mark.integration
     def test_proxy_authentication_in_url(self):
         """Test proxy URLs with authentication credentials."""
         config = JiraConfig(
@@ -154,7 +149,6 @@ class TestProxyConfigurationEnhanced(BaseAuthTest):
         assert "proxyuser:proxypass" in config.http_proxy
         assert "proxyuser:proxypass" in config.https_proxy
 
-    @pytest.mark.integration
     def test_socks_proxy_configuration(self, monkeypatch):
         """Test SOCKS proxy configuration for both services."""
         mock_jira = MagicMock()
@@ -185,7 +179,6 @@ class TestProxyConfigurationEnhanced(BaseAuthTest):
             == "socks5://socksuser:sockspass@socks.company.com:1080"
         )
 
-    @pytest.mark.integration
     def test_proxy_bypass_for_internal_domains(self, monkeypatch):
         """Test that requests to NO_PROXY domains bypass the proxy."""
         # Set up environment
@@ -204,7 +197,6 @@ class TestProxyConfigurationEnhanced(BaseAuthTest):
         assert os.environ["NO_PROXY"] == "*.internal.com,localhost,127.0.0.1"
         assert "internal.com" in config.no_proxy
 
-    @pytest.mark.integration
     def test_proxy_error_handling(self, monkeypatch):
         """Test proper error handling when proxy connection fails."""
         # Mock to simulate proxy connection failure
@@ -224,7 +216,6 @@ class TestProxyConfigurationEnhanced(BaseAuthTest):
         with pytest.raises(ProxyError, match="Unable to connect to proxy"):
             JiraClient(config=config)
 
-    @pytest.mark.integration
     def test_proxy_configuration_precedence(self):
         """Test that explicit proxy config takes precedence over environment."""
         with patch.dict(
@@ -247,7 +238,6 @@ class TestProxyConfigurationEnhanced(BaseAuthTest):
             assert config.http_proxy == "http://explicit.proxy.com:8080"
             assert config.https_proxy == "https://explicit.proxy.com:8443"
 
-    @pytest.mark.integration
     def test_mixed_proxy_and_ssl_configuration(self, monkeypatch):
         """Test proxy configuration works correctly with SSL verification disabled."""
         mock_confluence = MagicMock()
@@ -284,7 +274,6 @@ class TestProxyConfigurationEnhanced(BaseAuthTest):
         assert mock_session.proxies["http"] == "http://proxy.company.com:8080"
         assert config.ssl_verify is False
 
-    @pytest.mark.integration
     def test_proxy_with_oauth_configuration(self):
         """Test proxy configuration works with OAuth authentication."""
         with MockEnvironment.oauth_env() as env_vars:
